+++ Source/Scene/Vector3DTileContent.js
@@ -494,16 +494,16 @@ define([
         }
     };
 
-    Vector3DTileContent.prototype.applyStyle = function(frameState, style) {
+    Vector3DTileContent.prototype.applyStyle = function(style) {
         createFeatures(this);
         if (defined(this._polygons)) {
-            this._polygons.applyStyle(frameState, style, this._features);
+            this._polygons.applyStyle(style, this._features);
         }
         if (defined(this._polylines)) {
-            this._polylines.applyStyle(frameState, style, this._features);
+            this._polylines.applyStyle(style, this._features);
         }
         if (defined(this._points)) {
-            this._points.applyStyle(frameState, style, this._features);
+            this._points.applyStyle(style, this._features);
         }
     };